Move `getVehicleType` to own file
